requirements.txt
beautifulsoup4==4.6.0
breadability==0.1.20
certifi==2017.11.5
chardet==3.0.4
cssselect==1.0.1
docopt==0.6.2
feedfinder2==0.0.4
feedparser==5.2.1
flake8==3.5.0
idna==2.6
jieba3k==0.35.1
lxml==4.1.1
mccabe==0.6.1
newspaper3k==0.2.5
nltk==3.2.5
numpy==1.13.3
olefile==0.44
pew==1.1.2
Pillow==4.3.0
pipenv==9.0.1
psutil==5.3.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
requests-file==1.4.2
six==1.11.0
sumy==0.7.0
tldextract==2.2.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
